+++ backend/tests/Functional/Controller/App/GroupControllerTest.php
@@ -101,7 +101,8 @@ class GroupControllerTest extends WebTestCase
         $this->assertSame($body1, $body2);
 
         $this->assertSame([
-            ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE]
+            ['id' => $this->group1Id, 'name' => 'g1',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]
         ], $body1);
     }
 
@@ -139,7 +140,8 @@ class GroupControllerTest extends WebTestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertSame([
-            ['id' => $this->group0Id, 'name' => 'g0', 'visibility' => Group::VISIBILITY_PRIVATE],
+            ['id' => $this->group0Id, 'name' => 'g0',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
         ], $this->parseJsonBody($response));
     }
 
@@ -182,7 +184,8 @@ class GroupControllerTest extends WebTestCase
         $expected = [[
             'character' => ['id' => 123, 'name' => 'C1', 'corporation' => null],
             'groups' => [
-                ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE]
+                ['id' => $this->group1Id, 'name' => 'g1',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]
             ],
         ], [
             'character' => ['id' => 789, 'name' => 'C3', 'corporation' => [
@@ -191,8 +194,10 @@ class GroupControllerTest extends WebTestCase
                 ]
             ]],
             'groups' => [
-                ['id' => $this->group0Id, 'name' => 'g0', 'visibility' => Group::VISIBILITY_PRIVATE],
-                ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE],
+                ['id' => $this->group0Id, 'name' => 'g0',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+                ['id' => $this->group1Id, 'name' => 'g1',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
             ],
         ]];
         $this->assertSame($expected, $body);
@@ -224,7 +229,8 @@ class GroupControllerTest extends WebTestCase
         $expected = [[
             'character' => ['id' => 123, 'name' => 'C1', 'corporation' => null],
             'groups' => [
-                ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE]
+                ['id' => $this->group1Id, 'name' => 'g1',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]
             ],
         ], [
             'character' => ['id' => 789, 'name' => 'C3', 'corporation' => [
@@ -293,8 +299,10 @@ class GroupControllerTest extends WebTestCase
         $this->assertEquals(200, $response->getStatusCode());
 
         $this->assertSame([
-            ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE],
-            ['id' => $this->group4Id, 'name' => 'g4', 'visibility' => Group::VISIBILITY_PRIVATE],
+            ['id' => $this->group1Id, 'name' => 'g1',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+            ['id' => $this->group4Id, 'name' => 'g4',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
         ], $this->parseJsonBody($response));
     }
 
@@ -336,13 +344,17 @@ class GroupControllerTest extends WebTestCase
 
         $expected = [[
             'id' => 500, 'name' => 'five', 'ticker' => '-5-', 'groups' => [
-                ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE],
-                ['id' => $this->group4Id, 'name' => 'g4', 'visibility' => Group::VISIBILITY_PRIVATE],
+                ['id' => $this->group1Id, 'name' => 'g1',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+                ['id' => $this->group4Id, 'name' => 'g4',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
             ]
         ], [
             'id' => 501, 'name' => 'f1', 'ticker' => '-51-', 'groups' => [
-                ['id' => $this->group0Id, 'name' => 'g0', 'visibility' => Group::VISIBILITY_PRIVATE],
-                ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE],
+                ['id' => $this->group0Id, 'name' => 'g0',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+                ['id' => $this->group1Id, 'name' => 'g1',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
             ]
         ]];
         $this->assertSame($expected, $body);
@@ -404,8 +416,10 @@ class GroupControllerTest extends WebTestCase
         $this->assertEquals(200, $response->getStatusCode());
 
         $this->assertSame([
-            ['id' => $this->group4Id, 'name' => 'g4', 'visibility' => Group::VISIBILITY_PRIVATE],
-            ['id' => $this->group5Id, 'name' => 'g5', 'visibility' => Group::VISIBILITY_PRIVATE],
+            ['id' => $this->group4Id, 'name' => 'g4',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+            ['id' => $this->group5Id, 'name' => 'g5',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
         ], $this->parseJsonBody($response));
     }
 
@@ -447,13 +461,17 @@ class GroupControllerTest extends WebTestCase
 
         $expected = [[
             'id' => 100, 'name' => 'one', 'ticker' => '-1-', 'groups' => [
-                ['id' => $this->group4Id, 'name' => 'g4', 'visibility' => Group::VISIBILITY_PRIVATE],
-                ['id' => $this->group5Id, 'name' => 'g5', 'visibility' => Group::VISIBILITY_PRIVATE],
+                ['id' => $this->group4Id, 'name' => 'g4',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+                ['id' => $this->group5Id, 'name' => 'g5',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
             ]
         ], [
             'id' => 101, 'name' => 'o1', 'ticker' => '-11-', 'groups' => [
-                ['id' => $this->group0Id, 'name' => 'g0', 'visibility' => Group::VISIBILITY_PRIVATE],
-                ['id' => $this->group4Id, 'name' => 'g4', 'visibility' => Group::VISIBILITY_PRIVATE],
+                ['id' => $this->group0Id, 'name' => 'g0',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+                ['id' => $this->group4Id, 'name' => 'g4',
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
             ]
         ]];
         $this->assertSame($expected, $body);
@@ -490,7 +508,8 @@ class GroupControllerTest extends WebTestCase
 
         $body = $this->parseJsonBody($response);
         $this->assertSame([
-            ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE],
+            ['id' => $this->group1Id, 'name' => 'g1',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
         ], $body);
     }
 
@@ -514,9 +533,12 @@ class GroupControllerTest extends WebTestCase
 
         $body = $this->parseJsonBody($response);
         $this->assertSame([
-            ['id' => $this->group1Id, 'name' => 'g1', 'visibility' => Group::VISIBILITY_PRIVATE],
-            ['id' => $this->group4Id, 'name' => 'g4', 'visibility' => Group::VISIBILITY_PRIVATE],
-            ['id' => $this->group5Id, 'name' => 'g5', 'visibility' => Group::VISIBILITY_PRIVATE],
+            ['id' => $this->group1Id, 'name' => 'g1',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+            ['id' => $this->group4Id, 'name' => 'g4',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+            ['id' => $this->group5Id, 'name' => 'g5',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
         ], $body);
     }