+++ backend/tests/Functional/Controller/User/CorporationControllerTest.php
@@ -152,11 +152,14 @@ class CorporationControllerTest extends WebTestCase
         $this->assertSame(
             [
                 ['id' => 111, 'name' => 'corp 1', 'ticker' => 't1', 'alliance' => null, 'groups' => [
-                    ['id' => $this->group1->getId(), 'name' => 'group 1', 'visibility' => Group::VISIBILITY_PRIVATE]
+                    ['id' => $this->group1->getId(), 'name' => 'group 1',
+                        'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]
                 ]],
                 ['id' => 222, 'name' => 'corp 2', 'ticker' => 't2', 'alliance' => null, 'groups' => [
-                    ['id' => $this->group1->getId(), 'name' => 'group 1', 'visibility' => Group::VISIBILITY_PRIVATE],
-                    ['id' => $this->gid2, 'name' => 'group 2', 'visibility' => Group::VISIBILITY_PRIVATE]
+                    ['id' => $this->group1->getId(), 'name' => 'group 1',
+                        'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
+                    ['id' => $this->gid2, 'name' => 'group 2',
+                        'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]
                 ]]
             ],
             $this->parseJsonBody($response)
@@ -514,7 +517,8 @@ class CorporationControllerTest extends WebTestCase
         $response = $this->runApp('GET', '/api/user/corporation/222/get-groups-tracking');
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertSame(
-            [['id' => $this->group1->getId(), 'name' => 'group 1', 'visibility' => Group::VISIBILITY_PRIVATE]],
+            [['id' => $this->group1->getId(), 'name' => 'group 1',
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]],
             $this->parseJsonBody($response)
         );
     }