+++ backend/tests/Functional/Controller/User/WatchlistControllerTest.php
@@ -908,6 +908,7 @@ class WatchlistControllerTest extends WebTestCase
             'id' => $this->group1->getId(),
             'name' => 'g1',
             'visibility' => Group::VISIBILITY_PRIVATE,
+            'autoAccept' => false,
         ]], $this->parseJsonBody($response));
     }
 
@@ -1013,6 +1014,7 @@ class WatchlistControllerTest extends WebTestCase
             'id' => $this->group3->getId(),
             'name' => 'g3',
             'visibility' => Group::VISIBILITY_PRIVATE,
+            'autoAccept' => false,
         ]], $this->parseJsonBody($response));
     }