+++ backend/tests/Unit/Entity/GroupTest.php
@@ -20,7 +20,7 @@ class GroupTest extends TestCase
         $group->setName('g.name');
 
         $this->assertSame(
-            ['id' => null, 'name' => 'g.name', 'visibility' => Group::VISIBILITY_PRIVATE],
+            ['id' => null, 'name' => 'g.name', 'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false],
             json_decode((string) json_encode($group), true)
         );
     }
@@ -52,6 +52,14 @@ class GroupTest extends TestCase
         $this->assertsame(Group::VISIBILITY_PRIVATE, $group->getVisibility());
         $group->setVisibility(Group::VISIBILITY_PUBLIC);
         $this->assertsame(Group::VISIBILITY_PUBLIC, $group->getVisibility());
+    }
+
+    public function testSetGetAutoAccept()
+    {
+        $char = new Group();
+        $this->assertFalse($char->getAutoAccept());
+        $char->setAutoAccept(true);
+        $this->assertTrue($char->getAutoAccept());
     }
 
     public function testAddGetRemoveApplication()